+++ packages/env/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Bug fix
+-   PHP 7.3 and 7.4 must use PHPUnit 9.
+
+### Enhancement
+-   It's now possible to run PHPUnit tests on PHP 8.1 and 8.2.
+
 ## 5.9.0 (2023-01-02)
 
 ## 5.8.0 (2022-12-14)